generate tables

+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('queue')->index();
             $table->longText('payload');
             $table->unsignedTinyInteger('attempts');
-            $table->unsignedInteger('reserved_at')->nullable();
-            $table->unsignedInteger('available_at');
-            $table->unsignedInteger('created_at');
+            $table->unsignedBigInteger('reserved_at')->nullable();
+            $table->unsignedBigInteger('available_at');
+            $table->unsignedBigInteger('created_at');
         });
 
         Schema::create('job_batches', function (Blueprint $table) {
